backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
langchain==0.3.18,<1.0.0
langchain-community==0.3.17
llama-cpp-python==0.3.7
langchain-openai==0.3.5
python-dotenv==1.0.1
sentence-transformers==3.4.1
faiss-cpu==1.10.0  # Use faiss-gpu if GPU is available 
# MongoDB requirements
motor==3.7.0        # MongoDB async Python driver
pymongo==4.11.1     # MongoDB Python driver
pydantic==2.10.6     # For data validation